refactor(uilanding): extract capture, latency, and impairment proxy handlers with tests
Extract 13 handler classes from the monolithic uilanding.py into three
dedicated modules under src/handlers/:

- handlers/capture.py: CaptureWebSocketHandler, CaptureBridgesAPIHandler,
  CaptureStatusAPIHandler, CaptureStartAPIHandler, CaptureStopAPIHandler.
  Includes a standalone _get_topo_build_data() that reads ACCESS_INFO.yaml
  directly, removing the circular import dependency on uilanding.py.

- handlers/latency.py: LatencyBridgesAPIHandler, LatencyEnableAPIHandler,
  LatencyDisableAPIHandler, LatencyDisableAllAPIHandler.

- handlers/impairments.py: ImpairmentsBridgesAPIHandler,
  ImpairmentsConfigureAPIHandler, ImpairmentsClearAPIHandler,
  ImpairmentsClearAllAPIHandler.

All handlers use module-level tornado.httpclient imports and the same
primary/fallback URL pattern (host.docker.internal:8089 -> 172.17.0.1:8089).

uilanding.py updated to import from the new modules and removes the
~850-line inline definitions; route registrations are unchanged.

Adds 16 new tests across test_capture.py, test_latency.py, and
test_impairments.py covering proxy pass-through, primary/fallback
failover, 503 on total failure, auth enforcement, and topology enrichment.
All 224 tests pass.
